Update README.md
Removing the "Scope" that, with current formatting, appears at the end of the preceding line, creating the confusing statement:
> Explanation of how this impacts alignment between GA4GH standards Scope
after rendering the markdown.

Given the brevity of this line by itself and the lack of any formatting, I think this was an artifact from the construction of this document and propose removal. If this was intended to be in there, a little more clarity on "Scope" should be provided, as I feel that the other bullets here subsume issue scoping.